requirements.txt
numpy==2.1.2
Pillow==11.0.0
Requests==2.32.3
torch==2.4.1
